requirements.txt
numpy==1.19.1
tqdm==4.48.2
matplotlib==3.1.2
torchvision==0.5.0+cu100
pytorch3d==0.2.5
torch==1.4.0+cu100
Pillow==8.0.1
